Function playRound to play one round of the game
Added function that allows the user to play one round of the game and
logs a result string such as "You win! Rock beats paper." to the
console. The playRound function takes getHumanChoice and
getComputerChoice as arguments.
Created two variables humanSelection and computerSelection to hold the
return values of getHumanChoice and getComputerChoice respectively.
Both variables are then passed to the playRound function as
arguments